modèle sur l'attaque, normalement le package combat est fait

+++ code/app/src/main/java/com/example/ultimateandroid/modele/combat/CalculCoefficient.java
@@ -2,6 +2,16 @@ package com.example.ultimateandroid.modele.combat;
 
 import com.example.ultimateandroid.modele.pokemon.Type;
 
+/**
+ * Interface fonctionnelle gérant le calcul du coefficient entre les types
+ */
 public interface CalculCoefficient {
-    public int getCoefficient(Type typeAttaquant, Type typeAttaque);
+
+    /**
+     * Retourne un coefficiant de dégâts en fonction du type de l'attaque lancée par rapport au type du pokemon attaqué
+     * @param typeAttaquant : type de l'attaque utilisée
+     * @param typeAttaque : type du pokemon attaqué
+     * @return
+     */
+    public float getCoefficient(Type typeAttaquant, Type typeAttaque);
 }
